Develop

+++ app/Filament/Clusters/Settings/Resources/Accounts/Pages/EditAccount.php
@@ -22,6 +22,9 @@ class EditAccount extends EditRecord
             LocaleSwitcher::make(),
             DeleteAction::make()
                 ->action(function (Model $record) {
+                    if (! $record instanceof \App\Models\Account) {
+                        throw new \Exception('Invalid record type');
+                    }
                     $accountService = new AccountService;
                     $accountService->delete($record);
                     $this->redirect(AccountResource::getUrl('index'));
@@ -31,6 +34,10 @@ class EditAccount extends EditRecord
 
     protected function handleRecordUpdate(Model $record, array $data): Model
     {
+        if (! $record instanceof \App\Models\Account) {
+            throw new \Exception('Invalid record type');
+        }
+
         $accountService = new AccountService;
 
         return $accountService->update($record, $data);
